pyproject.toml
[project]
name = "django-moses"
version = "0.11.7"

[tool.poetry]
name = "django-moses"
description = "Advanced authentication with OTP and phone number verification"
version = "0.11.7"
authors = ["Vassily Vorobyov <vassily.v.v@gmail.com>"]
packages = [{ include = "moses/**/*", from = "." }]

[tool.poetry.dependencies]
python = "^3.13"
django = "^4.2.17"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.3.1"
pyotp = "^2.8.0"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
factory_boy = "^3.2.1"

[tool.poetry.group.dev.dependencies]
psycopg = "^3.1.8"

[tool.setuptools]
packages = ['moses', 'moses.migrations', 'moses.views']

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_project.settings"
pythonpath = "."
python_files = ["moses/**/*.py", "test_project/*.py"]

